requirements/base.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2024.1
sqlparse==0.5.3
